Reorganize EditorToolbar and its components
