+++ packages/online-editor/src/editor/EditorPage.tsx
@@ -18,7 +18,7 @@ import * as React from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useHistory } from "react-router";
 import { useRoutes } from "../navigation/Hooks";
-import { EditorToolbar } from "./EditorToolbar";
+import { EditorToolbar } from "./Toolbar/EditorToolbar";
 import { useDmnTour } from "../tour";
 import { useOnlineI18n } from "../i18n";
 import { ChannelType } from "@kie-tools-core/editor/dist/api";
