+++ packages/online-editor/src/editor/Toolbar/KebabDropdown.tsx
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from "react";
+import { Dropdown, DropdownPosition, DropdownToggle } from "@patternfly/react-core/dist/js/components/Dropdown";
+import EllipsisVIcon from "@patternfly/react-icons/dist/js/icons/ellipsis-v-icon";
+
+export function KebabDropdown(props: {
+  id: string;
+  items: React.ReactNode[];
+  state: [boolean, React.Dispatch<React.SetStateAction<boolean>>];
+}) {
+  return (
+    <Dropdown
+      className={"kie-tools--masthead-hoverable"}
+      isOpen={props.state[0]}
+      isPlain={true}
+      position={DropdownPosition.right}
+      onSelect={() => props.state[1](false)}
+      toggle={
+        <DropdownToggle
+          id={props.id}
+          toggleIndicator={null}
+          onToggle={(isOpen) => props.state[1](isOpen)}
+          ouiaId={props.id}
+        >
+          <EllipsisVIcon />
+        </DropdownToggle>
+      }
+      dropdownItems={props.items}
+    />
+  );
+}
